# docs-site/public/openapi.yaml
openapi: 3.1.0
info:
  title: clomp API
  version: 0.2.0
  description: >-
    Tamper-evident audit trail for security work. Events are append-only and
    hash-chained; exports verify offline. See the hash chain specification in
    the documentation for the exact hashing rules.
  license:
    name: MIT
    url: https://github.com/sockulags/clomp/blob/main/LICENSE
...
servers:
  - url: https://clomp.example.com
    description: Your self-hosted installation
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Machine credential (clomp_live_…), scoped to its tenant.
    session:
      type: apiKey
      in: cookie
      name: clomp_session
      description: Browser session obtained via POST /api/auth/login.
...
security:
  - apiKey: []
  - session: []
paths:
...
  /api/export/jsonl:
    get:
      tags: [export]
      summary: Offline-verifiable JSONL export
      description: One JSON object per line — events then signed checkpoints. Verify with `clomp verify-file`.
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: application/x-ndjson stream
          content:
            application/x-ndjson: {}
  /api/export/report:
    get:
      tags: [export]
      summary: Audit-ready PDF report
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: framework, in: query, schema: { type: string, enum: [soc2, nis2] }, description: Limit framework mappings to one framework }
      responses:
        '200':
          description: PDF
          content:
            application/pdf: {}
...